Reorganize infer.py into infer/
