+++ myia/infer/__init__.py
@@ -1,0 +1,26 @@
+"""Inference engine (types, values, etc.)."""
+
+from .core import (  # noqa
+    InferenceLoop,
+    EvaluationCache,
+)
+
+from .graph_infer import (  # noqa
+    Context,
+    Track,
+    InferenceEngine,
+    Reference,
+    Inferrer,
+    GraphInferrer,
+    PrimitiveInferrer,
+    PartialInferrer,
+    register_inferrer,
+)
+
+from .utils import (  # noqa
+    ANYTHING,
+    InferenceError,
+    MyiaTypeError,
+    MyiaShapeError,
+    ValueWrapper,
+)
