+++ myia/infer/graph_infer.py
@@ -1,48 +1,14 @@
-"""Inference engine (types, values, etc.)."""
+"""Inference engine for Myia graphs."""
 
 import asyncio
-from heapq import heappush, heappop
 from types import FunctionType
 
-from .dtype import Type, Function
-from .ir import is_constant, is_constant_graph, is_apply
-from .utils import Named, Partializable, UNKNOWN
+from ..dtype import Type, Function
+from ..ir import is_constant, is_constant_graph, is_apply
+from ..utils import Partializable, UNKNOWN
 
-
-# Represents an unknown value
-ANYTHING = Named('ANYTHING')
-
-
-class InferenceError(Exception):
-    """Inference error in a Myia program.
-
-    Attributes:
-        message: The error message.
-        refs: A list of references which are involved in the error,
-            e.g. because they have the wrong type or don't match
-            each other.
-        traceback_refs: A map from a context to the first reference in
-            that context that fails to resolve because of this error.
-            This represents a traceback of sorts.
-
-    """
-
-    def __init__(self, message, refs=[], app=None):
-        """Initialize an InferenceError."""
-        super().__init__(message, refs)
-        self.message = message
-        self.refs = refs
-        self.traceback_refs = {}
-        if app is not None:
-            self.traceback_refs[app.context] = app
-
-
-class MyiaTypeError(InferenceError):
-    """Type error in a Myia program."""
-
-
-class MyiaShapeError(InferenceError):
-    """Shape error in a Myia program."""
+from .core import InferenceLoop, EvaluationCache
+from .utils import ANYTHING, InferenceError, MyiaTypeError
 
 
 def type_error_nargs(ident, expected, got):
@@ -51,29 +17,6 @@ def type_error_nargs(ident, expected, got):
         f'Wrong number of arguments for {ident}:'
         f' expected {expected}, got {got}.'
     )
-
-
-class ValueWrapper:
-    """Wrapper for an inferred value.
-
-    Values may be wrapped using subclasses of ValueWrapper, associating them
-    with tracking data or metadata.
-    """
-
-    def __init__(self, value):
-        """Initialize a ValueWrapper."""
-        self.value = value
-
-    def __hash__(self):
-        return hash(self.value)
-
-    def __eq__(self, other):
-        return type(other) is type(self) \
-            and self.value == other.value
-
-    @property
-    def __unwrapped__(self):
-        return self.value
 
 
 #########
@@ -509,94 +452,6 @@ class VirtualReference:
 ########
 
 
-class _TodoEntry:
-    def __init__(self, order, handler):
-        self.order = order
-        self.handler = handler
-
-    def __lt__(self, other):
-        return self.order < other.order
-
-
-class _InferenceLoop(asyncio.AbstractEventLoop):
-    """EventLoop implementation for use with the inferrer.
-
-    This event loop doesn't allow scheduling tasks and callbacks with
-    `call_later` or `call_at`, which means the `timeout` argument to methods
-    like `wait` will not work. `run_forever` will stop when it has exhausted
-    all work there is to be done. This means `run_until_complete` may finish
-    before it can evaluate the future, which suggests an infinite loop.
-    """
-
-    def __init__(self, debug=False):
-        self._running = False
-        self._todo = []
-        self._debug = debug
-        self._futures = []
-        self._errors = []
-
-    def get_debug(self):
-        return self._debug
-
-    def run_forever(self):
-        self._running = True
-        while self._todo and self._running:
-            todo = heappop(self._todo)
-            h = todo.handler
-            if isinstance(h, asyncio.Handle):
-                h._run()
-            else:
-                fut = asyncio.ensure_future(h, loop=self)
-                self._futures.append(fut)
-
-    def is_running(self):
-        return self._running
-
-    def is_closed(self):
-        return not self.is_running()
-
-    def schedule(self, x, order=0):
-        heappush(self._todo, _TodoEntry(order, x))
-
-    def collect_errors(self):
-        futs, self._futures = self._futures, []
-        errors, self._errors = self._errors, []
-        for fut in futs:
-            if fut.done():
-                exc = fut.exception()
-            else:
-                exc = InferenceError(
-                    f'Could not run inference to completion.'
-                    ' There might be an infinite loop in the program'
-                    ' which prevents type inference from working.',
-                    refs=[]
-                )
-            if exc is not None:
-                errors.append(exc)
-        return errors
-
-    def call_soon(self, callback, *args, context=None):
-        h = asyncio.Handle(callback, args, self)
-        heappush(self._todo, _TodoEntry(0, h))
-        return h
-
-    def call_later(self, delay, callback, *args, context=None):
-        raise NotImplementedError(
-            '_InferenceLoop does not allow timeouts or time-based scheduling.'
-        )
-
-    def call_at(self, when, callback, *args, context=None):
-        raise NotImplementedError(
-            '_InferenceLoop does not allow time-based scheduling.'
-        )
-
-    def create_task(self, coro):
-        return asyncio.Task(coro, loop=self)
-
-    def create_future(self):
-        return asyncio.Future(loop=self)
-
-
 class InferrerEquivalenceClass:
     """An equivalence class between a set of Inferrers.
 
@@ -752,42 +607,6 @@ class EquivalencePool:
             self.engine.schedule(self.check(), order=1000)
 
 
-class EvaluationCache:
-    """Key/value store where keys are associated to Futures.
-
-    Attributes:
-        cache: The cache.
-        loop: The InferenceLoop for async evaluation.
-        keycalc: An async function that takes a key and returns
-            the value associated to that key.
-    """
-
-    def __init__(self, loop, keycalc):
-        """Initialize an EvaluationCache."""
-        self.cache = {}
-        self.loop = loop
-        self.keycalc = keycalc
-
-    def get(self, key):
-        """Get the future associated to the key."""
-        if key not in self.cache:
-            self.set(key, self.keycalc(key))
-        return self.cache[key]
-
-    def set(self, key, coro):
-        """Associate a key to a coroutine."""
-        self.cache[key] = self.loop.create_task(coro)
-
-    def set_value(self, key, value):
-        """Associate a key to a value.
-
-        This will wrap the value in a Future.
-        """
-        fut = asyncio.Future(loop=self.loop)
-        fut.set_result(value)
-        self.cache[key] = fut
-
-
 class InferenceEngine:
     """Infer various properties about nodes in graphs.
 
@@ -813,7 +632,7 @@ class InferenceEngine:
                  required_tracks=None,
                  eq_class=EquivalencePool):
         """Initialize the InferenceEngine."""
-        self.loop = _InferenceLoop()
+        self.loop = InferenceLoop()
         self.pipeline = pipeline
 
         self.graph = graph
